refactor: extract unicode-ellipsis into its own crate
